+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+### 0.0.7 (2020-02-06)
+* Nested models/entities getters/setters updated.
+
+
+### 0.0.6 (2020-02-06)
+* Nested models/entities support added.
+
+
 ### 0.0.5 (2020-01-31)
 * Example page added.
 * Example model added.